requirements.txt
streamlit
torch
torchvision
facenet-pytorch
Pillow
gdown